requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
openai==1.12.0
pandas==2.2.0
resend==0.7.0
reportlab==4.0.9
celery==5.3.6
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
